Try to fit things a bit better on the page.
Now the margins aren't such that content flows off the top,
but there is still a lot of white space at the bottom of each page.

I believe this is due to how I am scaling the entire body to get
it to be narrow enough to fit without lots of wrapping.

I believe a better solution would be to keep body width at 100% and
mess with font sizes and such instead, but I don't have time for that
right now.